Refactor app shell to server layout with client islands
