+++ app/HomeClientEffects.tsx
@@ -1,0 +1,8 @@
+"use client";
+
+import useSocket from "../hooks/useSocket";
+
+export default function HomeClientEffects() {
+  useSocket();
+  return null;
+}
