+++ components/MobileTocButton.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import React, { useEffect, useState } from "react";
 import TOC from "./TOC";
